force gradient w.r.t. model param

+++ geomatry/ff/springs.py
@@ -49,15 +49,15 @@ class SpringFF(baseFF):
 
 
 def _random_spring_ff_param(max_Za: int) -> Tuple[FloatTensor, FloatTensor]:
-    k = torch.rand(max_Za + 1, max_Za + 1) / 2 # symmetric
+    k = torch.rand(max_Za + 1, max_Za + 1, dtype=torch.float64) / 2 # symmetric
     k = (k + k.T) / 2
-    r0 = torch.rand(max_Za + 1, max_Za + 1) / 2 # symmetric
+    r0 = torch.rand(max_Za + 1, max_Za + 1, dtype=torch.float64) / 2 # symmetric
     r0 = (r0 + r0.T) / 2
     return k, r0
 
 
 def _random_spring_system(N: int, N_pairs: int, max_Za: int, start_Za: int=0) -> Tuple[FloatTensor, IntTensor, IntTensor, IntTensor]:
-    Ra = torch.randn(N, 3)
+    Ra = torch.randn(N, 3, dtype=torch.float64)
     Ra.requires_grad_()
     Za = torch.randint(start_Za, max_Za + 1, (N,))
     full_combinations = list(combinations(range(N), 2))
